genai/requirements.txt
# -----------------------------------------
# Web Framework
Flask==3.1.0
Flask-Cors==6.0.0
gunicorn

# Authentication
PyJWT[crypto]==2.10.1
python-dotenv==1.0.1

# -----------------------------------------
# Database & ORM
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# -----------------------------------------
# LLM & LangChain Integration
#openai==1.30.3
langchain>=0.3.0
langchain-core>=0.3.0
openai>=1.3.5,<2

# -----------------------------------------
# Configuration & Utils
python-dotenv==1.0.1
requests==2.32.4

# -----------------------------------------
# FastAPI (not used if on Flask only)
# fastapi>=0.100.0
# uvicorn[standard]>=0.23.0
# pydantic>=2.0.0